Delete OrderPostTests, Add PostForAllTests
